python: Add a server listener that handles client message callbacks (#226)
This adds a server listener interface to the Python SDK.

This is implemented as a Protocol in python; pyo3 does not support
extending a Python class. To provide reasonable typing on
`start_server`, that's now a wrapping function in the public package, to
avoid the stub interface referencing the external type.

Since the ws-related code is starting to grow in python, I've moved
related classes and functions into `websocket_server.rs` (from lib.rs).
The exports are still in the top-level `foxglove` package for now.

To support the message data callback, this also adds support for the
`ClientPublish` capability and `supported_encodings`.

I updated the existing live_visualization example to exercise this.