python: support status publishing on server (#227)
This adds `publish_status` and `remove_status` to the Python server
interface, to match the Rust SDK.

I've also updated the existing server methods to not raise exceptions,
since the Rust SDK doesn't return resuts. Currently, the only way for
these to fail is if the server has been shut down.