Support parameter subscriptions (#238)
This builds on #231 to add parameter subscription support to the Python
SDK. The SDK takes care of all bookkeeping and this is purely
informational for an implementer.

This also removes "parametersSubscribe" from the public SDK. Because the
SDK manages the subscriptions itself, as long as the client advertises
the "parameters" capability, the SDK will automatically include the
former capability. This does mean that the server listener callbacks
would be called regardless, but both languages provide default no-op
implementations.