rust: Don't expose websocket service IDs (#242)
Websocket service IDs are an implementation detail to accommodate
`ws-protocol`. It's not something that our users need to be aware of.
Service names need to be unique, and so that should be the handle by
which a service is identified or removed. I realized this as I was
trying to implement the python plumbing for `remove_services`.

Now we need two maps for indexing services: one by name, one by ID. They
need to be synchronized, so I introduced a new `ServiceMap` struct.