rust: Wrap Duration and Timestamp schemas (#247)
This change provides a new wrapper implementation for the `Duration` and
`Timestamp` types, to align with the foxglove schemas for those things.

- Use `sec` and `nsec` as field names.
- Infallible translation into `prost_types` equivalents for serialization.
- Offer (feature-flagged) conversions from chrono types.
- Update python bindings.

Note that all of the "convenience" conversion traits (from `std::time`
and `chrono` types) are fallible, because the foxglove schema is more a
restrictive representation in each case. Unfortunately, this means folks
are writing `ts.try_into().unwrap()` instead of `ts.into()`. To cope
with that, we offer `SaturatingFrom` and `SaturatingInto` traits, which
allows users to opt-in to an infallible conversion with clearly-defined
behavior.

This change does not attempt to provide conversions for python types.
We'll deal with that in a future change.