Rust ws server connection graph support (#237)
Adds support for connection graphs to the websocket server. This follows
the design in the C++ ws-protocol server very closely.

Clients can subscribe to updates, which sends the full state of the
connection graph in response. [This is not
documented](https://github.com/foxglove/ws-protocol/blob/main/docs/spec.md#subscribe-connection-graph-updates)
, I think we should update the spec to document that, as it doesn't seem
to me like it would work without it. The [C++ server behaves this
way](https://github.com/foxglove/ws-protocol/blob/main/cpp/foxglove-websocket/include/foxglove/websocket/websocket_server.hpp#L1449).

Like with parameters, we trigger callbacks on the first subscriber and
last unsubscriber to save the user having to do their own bookkeeping.
We tell them when they need to start being concerned with the connection
graph, and when they can stop.

It adds a method to the server, and the server handles for updating a
connection graph. One passes a new, full connection graph in, and it
will compute the difference with the last known state and send that
difference to subscribers. Because it works with differences, that's why
we need to send the full state when a subscriber first subscribes to the
connection graph.

The idea is the client will probably maintain a ConnectionGraph object
in the ServerListener that tracks the current graph, and after updating
it will clone() it and pass it to publish_connection_graph. The server
holds onto that copy so it can compute the diff the next time
publish_connection_graph is called.