python: improve documentation (#254)
This improves the python documentation & example; notably:

- Enum variants are documented (also in the stubs; IDEs will pick this
up)
- ParameterValue nested classes are documented
- Added param types to docstrings
- All additional examples have been added to the docs
- The last remaining example was moved from sdk into sdk-examples

For the first two items, I resorted to manual documentation for the
reasons noted in index.rst. This is unfortunate, and I think we're going
to need to look into some sort of doc testing down the line, but it's
better than missing documentation today.