python: support asset serving (#259)
This adds support for an asset handler in the Python SDK. It follows the
general approach of services: we expect a single function that returns
the `bytes` of the asset, and call this function in a separate thread.

I also renamed "Request" -> "ServiceRequest" here. Even though the asset
request is only a single URI, I think this clarifies what the request is
for. It also localizes classes in the documentation.

I added a basic example which will serve "package://" assets from the
local directory. We could include URDF descriptions in the example
itself to make this easier to see out of the box; I haven't done so yet.

---------

Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>