python: Don't override message_schema for json schemas (#271)
I watched @amacneil today as he did this:

```python
chan = Channel("topic", message_encoding="msgpack", schema={ ... })
```

That doesn't work as expected, because we automatically override the
message_encoding to be json when you provide a schema that looks like a
dict jsonschema. Which is a silly thing to do, in retrospect. The schema
encoding does not imply a message encoding.