rust: Use Bytes instead of Vec<u8> in protobufs (#272)
For bytes fields on protobufs, we should use `Bytes` instead of
`Vec<u8>`. This enables zero-copy for buffers that are already `Bytes`
of course, but also for types `T: AsRef<[u8]> + Send + 'static` (via
`Bytes::from_owner`), such as you'd find for an `Arc<Vec<u8>>` or
`Mmap` or other various memory views.