python: organize websocket code into submodule (#276)
### Changelog

- [python] Move websocket-related classes into a `websocket` module

### Description

This re-organizes existing python code by introducing a `websocket`
module. `start_server` is still in the top-level namespace, along with
all other functions. I think we could go either way on this, but this
sort of follows the rust organization, which has the server at the top
level, with supporting classes in a websocket module.