Change ClientChannel schema data to binary; unify base64 handling (#285)
### Changelog
None

### Docs

None

### Description

The current websocket protocol uses JSON messages for advertisements (in
both directions), which requires binary schemas to be base64-encoded.
Currently the SDK handles base64-encoding when sending `protobuf`
schemas to the client (although we missed also doing this for
`flatbuffer` schemas), but ClientAdvertisements left the base64 handling
up to the user (they would just receive a String which may or may not be
base64-encoded). This PR adds base64 handling for the client
advertisements as well so the SDK handles both directions and doesn't
require the user to mess with base64.

- Changed `ClientChannel.schema` to `Option<Vec<u8>>`
- Added `JsonClientChannel` & `JsonClientAdvertise` types for
deserialization, and `ClientChannel::try_from(JsonClientChannel)` to do
the base64 decoding
- Treat `flatbuffer` as base64 as well as protobuf
([docs](https://docs.foxglove.dev/docs/connecting-to-data/frameworks/custom))