rust: Wrap Context in Arc<> (#280)
### Description
We're going to need to store references to the `Context` on both
channels (for access to subscriptions) and sinks (for dynamic
subscription management and unregister-on-drop), so we're going to want
an `Arc`. Since the context holds `Arc` references to channels and
sinks, channels and sinks will hold `Weak` references back to the
context.