rust: Add SinkId for identifying sinks (#290)
### Changelog
- rust: Add `SinkId` and new mandatory `Sink::id()` trait method.

### Description
We've been using Arc pointers to compare log sinks, which is… fine. But
it'd be a bit nicer if we just had a SinkId to use, like we have a
ChannelId. That would give us a natural way of indexing sinks, which
will be useful for both sink deregistration as well as subscription
management.