Support schemaless channels over websocket (#283)
This adds support for live visualization of JSON channels which don't
define a schema. Currently, the SDK requires a schema to be defined for
a websocket sink. A user can specify an empty schema, but that's a step
we shouldn't require them to take.

The ws server will behave the same for schemaless channels using message
encodings other than JSON, since Foxglove doesn't support this.