rust: Add subscription manager (#281)
### Changelog
None

### Description
This change adds a new type for tracking subscriptions of sinks to
channels, but doesn't make use of it yet. That will come in a later PR.

We're planning to support two kinds of subscriptions: static
subscriptions to all channels (e.g., for mcap), and dynamic
subscriptions to particular channels (e.g., for websockets). We maintain
these subscriptions as two separate sets. When subscriptions change, the
`Context` will be responsible for querying the subscription manager in
order to regenerate the `LogSinkSet` for affected channels.

Part of: FG-10743