rust: Manage websocket channel subscriptions dynamically (#293)
### Changelog
- Skip message serialization for a `TypedChannel` that has no subscribed
websocket clients.

### Description
This change makes use of the new subscription manager, to teach the
websocket to manage channel subscriptions dynamically based on client
demand. The websocket no longer automatically subscribes to all
channels. Instead, it only subscribes to channels that have one or more
subscribed clients.

To make that happen, we move the implementation of `Sink` from `Server`
to `ConnectedClient`. This is a bit simpler than the alternative, which
would involve maintaining an aggregation layer for propagating client
subscriptions back to the context. Anyway, it feels a bit more natural
to treat the clients as sinks.