rust: Remove arbitrary sleeps from tests (#298)
### Description
Now that we have websocket clients represented as individual sinks, we
can remove a couple arbitrary 100ms sleeps from our tests, and instead
periodically poll the number of subscribed sinks on a channel.

While we're at it, we might as well get rid of some other arbitrary
sleeps as well, by polling on `RecordingServerListener` state, or simply
waiting to receive messages on the websocket client.

I ran tests in a loop on my machine for ten minutes and didn't see any
failures.