rust: Fix websocket service schema encoding (#304)
### Description
When I implemented websocket services, I didn't know about the Service
Call panel in the UI, and so I did all my testing with a custom client
in rust. Turns out I botched the request/response schema encoding.

This change fixes the schema encoding, using the same logic that we use
for encoding channel schemas. The ws-services example is now compatible
with the UI.