rust: Set MSRV to 1.83.0 (#327)
### Changelog
- rust: Explicitly declare MSRV as 1.83.0

### Description
We currently use DebugTuple::finish_non_exhaustive,
which was introduced in rust 1.83.0. Explicitly declare that as the
MSRV, so that users get a nice error if their toolchain is too old.

If we wanted to support a lower MSRV, we could replace
`finish_non_exhaustive` with `finish` - it's not doing any heavy
lifting. If we do choose to do that, it looks like we can drop the MSRV
down to 1.80.0 (which is when `LazyLock` was stabilized).