rust: Add Channel::close (#334)
### Changelog
- rust: Add `Channel::close`

### Description
As part of making `Context` public, I realized that we probably don't
want to expose `Context::remove_channel_for_topic()`. We should really
be removing channels by ID, to avoid TOCTOU errors. We should expose
this functionality publicly as `Channel::close()`, to mirror the python
`Channel.close`.

While I was doing that, I realized that [my fix for python's
`Channel.close`](https://github.com/foxglove/foxglove-sdk/pull/329)
should've also removed a lie in the docstrings ("Destroying all
references to the channel will also close it") and completely missed
updating the generated typed channels. So this change fakes care of
those things too.

The python SDK issues a (debug-level) warning when attempting to log on
a channel that was explicitly closed. In this change, we move that
warning down to rust, so that it's common. With a few changes:
- The message is now issued at _warning_ level, because we don't want
users to miss it.
 - Add a per-channel warning throttler with a 5s interval.
- Handle the case where a channel is implicitly closed because the
context is dropped.
 - Python no longer has to wrap channels in Option<>.

Fixes: FG-11010
Fixes: FG-10877