Set library version with SDK language (#288)
### Changelog

The SDK language is now reflected in the `library` written to MCAP
files, and is included in `serverInfo` metadata under the key
`fg-library`.

### Description

Example: `foxglove-sdk-python/v0.3.0`.

The language name is set at build time, based on environment. For C++,
which is a wrapper around the C library, a function is called by the log
sinks when they're created to override that language.

The app may use the serverInfo metadata for troubleshooting or
performance analysis. Foxglove has some precedent for the "fg-" prefix
in user agents and HTTP header keys. "fg-library" echoes the "library"
header of mcap.