Add rust high level log! macro - type-safe logging without channels (#311)
Add channelless logging for rust via log! macro, similar to the Python log function.

The magic is in create_channel, which is called at most once for any
given log site. It creates the Channel or fetches an existing one from
the Context for the topic. It compares the schema and message_encoding
and panics if they differ, making it type-safe.

One catch is the Channel, once fetched or created, will live until the
end of the process. It can be added or removed from sinks or contexts,
but it can't ever be freed. It would be possible to provide a method to
free it, but I don't think that makes any sense.

The type of the log $msg for a given log site can't change in Rust
AFAIK. It's always the same type (even if that type were dynamic, e.g.
'Box<dyn Encode>' it's always a 'Box<dyn Encode>').

I added optional "named arguments" like keyword argument syntax in
Python for log_time, publish_time, sequence.

I didn't add a context argument, one because it would double the number
of cases in the macro, which are already many. But also because I think
it's not that useful. If you want to specify the context, it's better to
explicitly create the channel.

---------

Co-authored-by: Greg Smith <gasmith@foxglove.dev>