rust: Ensure feature-flagged items are documented (#344)
### Description
Turns out we need to provide docs.rs with configuration, if we want it to build
with --all-features so that non-default features get documented. Not so
surprising in retrospect.

Docs.rs generates docs with +nightly, which opens up features like
doc_auto_cfg, which is the handy thing that annotates feature-flagged
items as "available on crate feature <foo> only". To indicate that we're
running in a docs.rs-y environment, we pass `--cfg docsrs`. This idea is
lifted from tokio.

I also updated the CI script to run rustdoc with nightly and `--cfg
docsrs` so that we get coverage.