rust: Refactor websocket parameter subscriptions (#345)
### Description
As part of splitting up websocket.rs, I'm trying to disentangle the
`Server` from the `ConnectedClient`. Parameter subscriptions are one
case where the client and server are accessing each others' internal
state (and locks!).

It looks like we can simplify this by just moving parameter subscription
management to the server as a `HashMap<String, HashSet<ClientId>>`. The
client can be basically stateless. I think this also makes it easier to
reason about the correctness of the listener callbacks, by keeping the
logic and state in a central location.