rust: Refactor websocket connection graph management (#348)
### Description
This is another change towards separating the websocket server from the
connected client. Just as with parameters, we don't really need to
maintain state in two places, and the server is the natural location
because (a) the connection graph state is shared, and (b) the server
needs to track how many clients are subscribed in order to deliver
unsubscribe callbacks.

This change also fixes a bug: we weren't removing a client's graph
subscription on disconnect. I added unit test coverage for this case (c2
disconnects without sending unsubscribe).

It was convenient to move the list of subscribers into the
`ConnectionGraph` itself, so that it'd be under the server's mutex
without having to introduce a new struct. A little wasteful to have an
extra hashset in there, but it's not like this weird corner of the SDK
is going to get a lot of use anyway.