rust: Use new ws protocol implementation (#355)
### Description
In this change, we switch over to using the new ws protocol
implementation. The change is fairly mechanical, and shouldn't introduce
any functional changes.

I used the opportunity to split a few things from `websocket::protocol`
into `websocket::X`. These are basically just newtypes and conversions.

The tests shrank in size, perform stricter validation, and I find them a
little more readable. Not quite the eye-watering result I was hoping
for, but I think it's still a reasonable improvement.