rust: Add websocket connection poller (#360)
This should be the last step in separating the client from the server,
so they can move into different modules.

The idea is to encapsulate the poll loop, so that the server doesn't
need to care about its inner workings. It just calls an `async fn
Poller::run(self)` which consumes self and returns when the client
connection is closed.

While we're at it, we might as well move the whole `WebSocketStream`
into the poller, too. That way, we can easily ensure that the connection
is shutdown in the appropriate sequence - i.e., there are no races
between the poller's send loop and some external attempt to send a
close message. But to accomplish that, we need better signalling to
indicate _why_ the connection is being closed. For that, we introduce
a `ShutdownReason` enum and use a oneshot channel to deliver it
to the poll loop.

A surprising upshot of this rearrangement is that `Server::stop` is no
longer async, which has two big implications:
- We don't need `WebSocketBlockingHandle` anymore.
- We can implement `Drop` for `WebSocketServerHandle`, if we want to.