rust: Reject websocket clients after Server::stop (#361)
This fixes a bug in the websocket server, where the server may accept
new connections after `Server::stop()`, and fail to shut them down
properly.

The idea is to use `Server::clients` as the point of synchronization.
While stopping the server, we take the contents of that `CowVec` and
freeze it, so that new clients cannot be added. The concurrent accept
loop, if it is still running, may accept a connection, and the server
will handshake it, but then immediately shut it down after failing to
add it to `self.clients`.

This change also cancels the accept loop a little earlier, to reduce the
likelihood of a new connection being accepted once `Server::stop()` is
called. I also removed the compare_exchange on `self.started` - once a
server instance is stopped, it's stopped forever.