Remove publish_time and sequence from metadata options (#358)
This proposes to remove `publish_time` and `sequence` from `Metadata`,
which users can currently provide to `log()`.

### Motivation

#### Publish Time

`publish_time` is a message field in MCAP but does not have universal
support in the Foxglove app. Some panels do support this (e.g. plot),
but also support user-defined timestamp fields, where users are free to
define their semantics. Data imported to a Foxglove primary site is
ordered by log_time. The SDK should not include this only to support
writing to an MCAP field; the MCAP spec dictates that log_time should be
used if publish_time is unavailable.

#### Sequence

`sequence` is a message field in MCAP. It's useful for detecting
missing, out-of-order, or discontinuous messages. The existing
implementation is problematic because users are free to set sequence
numbers in some log calls, but not others, which would result in an
out-of-sequence stream with possible duplications. In ROS 2,
`std_msgs/Header` no longer includes a sequence
(https://design.ros2.org/articles/changes.html) because it was
inconsistent in ROS 1. Data streamed from a Foxglove primary site
produces its own sequence numbers.

### Notes

For the Rust SDK, where users work with metadata structs directly, I've
left `log_time` inside those structs. Arguably, it's less useful now
that there's only a single field — there's no longer a reason for
`::default`, and a future addition will likely mean manual changes for
those users. Still, I think the migration would be easier, and the
struct better communicates the intent.