python: fix default schema names; provide topic and schema getters (#368)
This improves the handling of schema definitions which aren't given a
name. Previously, we defaulted to a hard-coded string. This produces
warnings in the app if you create multiple channels with unnamed
schemas. Now, we'll generate based on a hash of the schema contents so
that differing schemas will have different names, with reasonable
probability. I used base64 encoding and prefixed with "schema-" so that
labeling within the app is clear to users.

This also exposes `topic` and `schema_name` on the Channel, and improves
the serialized representation of the channel class.