python: fix channel interface to provide ID (#370)
This adds `id()` to the Python channel interface. We publish
subscription notifications on the server callback interface with these
IDs, but they were not meaningful until now.

This also updates the typed channel interface with public getters for
schema and topic name to match the raw channel getters added in #368.