C/C++ error handling (#350)
Adds error returns and error details to the C APIs which can fail.

This is a rework of the error handling based on feedback from James.
Instead of returning error strings to C, we can return an error type
enum (which we just map directly from the enum members in FoxgloveError
in the crate root.)

The reason is we don't really need the full string errors. An error code
is enough, plus we log the full string message. The user can debug based
on the error code, and if that's not enough, enable the logs and check
them. It's not quite as convenient for debugging, but it's a lot more
convenient for library usage and we don't have to allocate and free
errors on the heap.

Additionally the C++ API was changed to modern C++ practice of using
Result-like return values instead of exceptions. This also means
constructors had to be replaced with static create methods instead, so
that it can return the result. For this I used tl::expected which is an
implementation of std::expected in C++ 23.
https://en.cppreference.com/w/cpp/utility/expected

One annoying thing is our error enum is duplicated as a C enum and then
duplicated again as a c++ enum. A bit of a pain to update correctly, so
I opened a ticket to generate these instead so they don't get out of
sync.

I added some tests to exercise the error handling in the C++ API.

One change I made is to remove the foxglove_server_free and
foxglove_mcap_free functions, preferring to use just the stop/close
methods. There's nothing one can do with a closed/stopped server or mcap
writer, so there's no point separating stop/close from free.