Support channels with a shared topic (#385)
### Changelog

Do not error or panic if topic names are re-used between incompatible
channels.

Breaking: some rust interfaces have changed

### Description

This makes a few adjustments to our handling of re-used topic names
within a context's channels.

At a high level, there's a bit of a mismatch between the handling of
channels and topics in Foxglove and MCAP. MCAP doesn't dictate anything
here. The app doesn't have an explicit concept of a Channel; it instead
relies on topic names. There's a slight complicating factor in that the
Rust MCAP library has a notion of a "canonical channel"; if you create a
second channel with identical topic, schema, etc., then you'll get back
a reference to the canonical channel. It appears to be the only MCAP
library to behave this way.

The motivation here (and initial bug report) is to ease playback of MCAP
data which was created from other sources, which may contain multiple
channels with the same topic name. I tried to take a pragmatic approach
here that may err on the side of being too flexible. Instead of
returning an error (or panicking), SDK will log a warning that a
duplicate topic name was encountered, and I updated documentation to
clarify that topic names *should* be unique for compatibility.

With duplicate topics in play, this brings live viz from the SDK more in
line with MCAP playback (in terms of app behavior).

#### Considerations / future enhancements

We could introduce something like a "strict mode" (opt-in) which does
panic when encountering state that won't be fully compatible with the
app. That could arguably make initial development or testing easier when
app compatibility is the goal.

We could consider adopting the "canonical channel" behavior for the WS
Server as well, in order to coalesce channels that share a schema.
Currently, the app will only subscribe to one channel associated with a
topic. We could instead investigate enhancements in the app for
subscriptions to shared topic names (and the equivalent for file
playback).

I did consider keeping the behavior stricter for now. (I don't think
panicking should be the default behavior in these cases, but I'm less
convinced that these shouldn't be errors.) But I think the approach here
is more practical in the near term.