python: Improved websocket parameters (#389)
### Changelog
- python: Improved websocket parameters

### Description
The purpose of this change is to make websocket parameters more
ergonomic in python, by implementing conversions to and from native
python types.

The basic idea is that we can infer the (type, value) from the type
information on a python object, so that the caller doesn't need to
specify it explicitly. Likewise, we can convert (type, value) back to a
single native python object using a `get_value()` method.

The implementation gets just a tiny bit hairy, because we need to
introduce some shim types to accomplish the conversions, but generally,
pyo3's trait system works in our favor.