Add context to C and C++ (#364)
This adds a new C function to create a context and free a context.

It adds a C++ wrapper around the context, and provides a way to pass it
when creating channels, mcap writers, and websocket servers. It now uses
shared_ptr under the hood, so it can be trivially copied (like the
Arc<Context> in Rust).

The mcap and websocketserver use options structs in their creation, so I
added a Context to each options struct as a way to set the context. And
set that to a default, null context, which gets replaced by the global
context on the Rust side. Channels don't use an options struct, so I
added an overloaded constructor with 4 params where the final one is the
context, again defaulting to a null Context.

Then I updated all the tests to use a context scoped to the test so they
don't pollute the global context anymore (yay!)