rust: Rework websocket responder payload types (#395)
### Changelog
- rust: Breaking: Changed websocket FetchAsset and Service handler types

### Description
I originally proposed we use `bytes::Bytes` for websocket responders,
which works fine for callback wrappers that need to return an
owned/static value, but it's too restrictive for the underlying
responder methods, which would be perfectly happy with any `T:
AsRef<[u8]>`.

And so, the core of this change is updating
`service::Responder::respond` and `AssetResponder::respond` to
accept a `T: AsRef<[u8]>`.

Since that change breaks compatibility (albeit in a mild way), we might
as well go a bit further:
- If we accept an `E: Display`, we can simplify several callers.
- Callback wrappers can return a `T: AsRef<[u8]>`.