python: Expose logging context (#399)
This makes `Context` available to the python SDK, and adds an example
for usage. To use context from python, a user explicitly passes a
context to a channel constructor, and also to the `open_mcap` or
`start_server` functions. The "create_channel" method is private, at
least for now, because of the way typed channels are implemented in the
Python SDK.

This also cleans up the behavior of the untyped channel class to remove
the restriction on duplicate topics. This brings it in line with typed
channels and with Rust.