rust: Fix log! macro to reuse matching channels (#400)
### Changelog
- rust: Fix log! macro to reuse matching channels

### Description
Commit f6087cb7 introduced a bug, which caused the `log!` macro would
unconditionally create a channel per callsite. This change fixes that
bug, by properly deduplicating channels when they are added to the
context.

I decided to extend this deduplicating behavior to all channels. I can't
think of any reason for someone to deliberately create multiple channels
with precisely the same attributes, that differ only by ID. This should
have no impact for MCAP files, and can only improve the websocket
experience.

While working on this change, I realized that the context's topic index
actually needs to hold a list of matching channels, so that we can (a)
search for duplicates and (b) handle removal properly. I also realized
that `Context::add_channel` doesn't need to return a `Result`, since it
cannot fail. And in turn, the `panic!()` in the `log!()` macro is
actually unreachable!

Fixes: FG-11405
Fixes: f6087cb7