C: Convince miri that FoxgloveStringBuf is safe (#416)
I was futzing with miri, to see if it could help us catch undefined
behavior in the C FFI. It seems fairly happy at the moment, though
I'd take that with a grain of salt, since our test coverage here is also
a little weak.

The one thing miri did notice is that the `FoxgloveStringBuf` was
casting a String to a const raw pointer, and then later casting to a
mutable raw pointer to recover it. This change helps miri to track the
pointer as a mutable reference.