Improve C++ documentation; rework docs publishing (#418)
This replaces the generated Doxygen HTML pages with Sphinx docs, using
the breathe & exhale extensions (BSD licensed). Doxygen is still used to
generate the XML source for these pages. I used a similar layout to the
python docs, with examples next to the API reference. I chose a
different theme, since I don't think any of these are 'standard' for
c++, and which I thought looked clean.

This also adds a new "Docs" workflow, which deploys both the C++ and
Python documentation to gh-pages. We build both c++ and python docs,
then upload in a single job. This has no dependencies on the deployment
job; we could introduce a callable workflow, but I don't think it's
needed.

I deployed C++ documentation from this branch:
https://foxglove.github.io/foxglove-sdk/cpp/