rust: make toplevel example snippet pastable into any function (#422)
### Changelog
None.
### Docs

Is a docs PR.
### Description

Today there is a code snippet showing the user how to use the foxglove
SDK, but to use it they must wrap it in a function that returns
`Result<T, E> where E: From<FoxgloveError>`. This means they'd need to
write non-trivial extra code to use this code as written. This PR
updates the example to use `expect`, which allows a user to get started
immediately from `cargo init .; cargo add foxglove`.

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->