Add `foxglove::Encode` derive macro (#324)
### Changelog
Add a `foxglove::Encode` derive macro to implement the `Encode` trait
for logging custom structs using a binary serialization format.

### Docs
None

### Description
When using the Rust SDK to log data, one might want to define
app-specific structs for logging. The current recommended approach for
this directs the user to derive the `Serialize` and `JsonSchema` traits
for their struct. The foxglove SDK then provides a specialization of the
`Encode` trait for structs which have implemented these two other
traits.

```rust
#[derive(serde::Serialize, schemars::JsonSchema)]
struct Message {
    msg: String,
    count: u64,
}
```

This has some downsides:
- Field names are repeated in every message (leading to message size
bloat for smaller messages and high-rate data)
- "byte" fields are serialized as base64 strings
- No support for u64 (larger integers are not properly deserialized in
the app)

This change introduces a new `foxglove::Encode` derive macro. The
purpose of this macro is to implement the `Encode` trait for a struct
and thereby making it possible to log the struct on a channel.

The macro generates an implementation that serializes the
struct using the protobuf serialization format.