Add foxglove schemas for C++ (#413)
This PR adds generated foxglove schema and channel types for C++. Tickets FG-11372 and FG-11373.

By having a dedicated channel type per schema type, we avoid the
overhead of dynamic dispatch (e.g. based on an enum with a giant switch
statement or virtual function table or similar).

It adds conversion code from C++ -> C. This may need to allocate in
cases where the C type has a different layout, so it also adds a
stack-based arena allocator used for those temporary allocations. The
arena is simply discarded when the function returns, as the converted
type doesn't outlive the function call.

The C -> Rust conversion code works the same way, but with lots of
unsafe code and ManuallyDrop to shoehorn borrowed data into owned types
without dropping it. Again a stack based arena is used (which is pinned
so it isn't accidentally moved). It should be possible to optimize this
in the future for types where the C layout of the protobuf struct could
be the same if we add repr(c) to it, we could avoid having to do this
conversion for those types (e.g. Color). The C++ code already does this
optimization, see isSameAsCType in generateSdkCpp.ts.

One gotcha is the Rust conversions can fail (e.g. invalid utf-8) so we
have to do all fallable conversions first, keeping the results inside
MnaullyDrop, before constructing the final ManuallyDrop<T> result.

We return the typed channel to C as a raw channel pointer, and then cast
it back to the appropriate typed channel in the log call.

The stack based allocators have a fixed buffer size of 128kb. They fall
back to heap allocation if they run out of space, tracking and freeing
the allocation when the arena is cleaned up. This seems extremely
unlikely as they're used only for small vectors and structs. I've
implemented many memory allocators in the past, but this is the first
time I've been able to include any in a PR at work!

The gist of this is we do some copying/transforming while other fields
are borrowed as-is (e.g. string, bytes, some vectors), but no heap
allocations and freeing on the logging path, apart from what is needed
to build the message schema object itself in C++.

I added a couple cherry-picked tests - it's not quite as good as
generating comprehensive tests, but this has dragged on for so long I
don't feel I have the luxury of time to do that.

Issue for more tests: FG-11383
Issue for example(s): FG-11374